\name{ranlip}
\alias{ranlip}

\title{Ranlip Package
}
\description{This function shows a list of function included in this toolbox
}
\usage{
  ranlip()
}

\details{The method of building the hat function, and generation of random variates using acceptance/ rejection described
}

\value{
  \item{output}{No return value}
}
\references{

  [1] G. Beliakov.  Class library ranlip for multivariate nonuniform random variate generation. Computer Physics Communications, 170:93-108, 2005.
  
  [2]  G.   Beliakov. Universal   nonuniform   random   vector   generator   based   on acceptance-rejection. ACM Transactions on Modeling and Computer Simulation, 15:205-232, 2005.
  
[3] L. Devroye Non-uniform Random Variate Generation. Springer Verlag New York, 1986.

[4] W.Hormann, A rejection technique for sampling from t-concave distribution. ACM Transactions on Mathematical Software, 21:182-193,1995.

[5]W.Hormann, J. Leydold and G Derflinger. Austomatic Nonuniform Random Variate Generation, Springer, Berlin, 2004.

[6] J. Leydold and W.Hormann. A sweep-plane algorithm for generatin random tuples in simple polytopes. Mathematics of Computation, 67:1617-1635, 1998.

[7] M. Luescher. A portable high-quality random number generator for lattice field theory calculation. Computer Physics Communications, 79:100-11, 1994.
}
\author{
    Gleb Beliakov, Daniela L. Calderon, Deakin University
}

\examples{
	ranlip()
}

\keyword{ranlip}
\keyword{ranlip}% __ONLY ONE__ keyword per line