\name{ranlip.Seed}
\alias{ranlip.Seed}

\title{Function of setting the seed
}
\description{Function for setting the seed of the default uniform random number generator ranlux.
}
\usage{
  ranlip.Seed(seed)
}

\arguments{
 \item{seed}{Integer value to seed the random generator}
}
\value{
  \item{output}{No return value, called to set the seed of the random generator.}
  
}
\author{
   Gleb Beliakov, Daniela L. Calderon, Deakin University
}

\examples{

	ranlip.Seed(17);
}

\keyword{ Seed}
\keyword{Seed}% __ONLY ONE__ keyword per line